+++ projects/fb-openlayers/src/lib/vessel/layer-vessel-trail.component.ts
@@ -123,26 +123,31 @@ export class VesselTrailComponent implements OnInit, OnDestroy, OnChanges {
 
   parseLocalTrail() {
     if(!this.localTrail) { return }
-    let c= fromLonLatArray(this.localTrail);
+    const c= fromLonLatArray(mapifyCoords(this.localTrail));
+    // let c= fromLonLatArray(this.localTrail);
     if(!this.trailLocal) { // create feature
-      this.trailLocal= new Feature( new LineString( mapifyCoords(c) ) );
+      this.trailLocal= new Feature( new LineString( c ) );
+      //this.trailLocal= new Feature( new LineString( mapifyCoords(c) ) );
       this.trailLocal.setId('trail.self.local');
       this.trailLocal.setStyle(this.buildStyle('local'));
     }
     else { //update feature
       this.trailLocal= this.source.getFeatureById('trail.self.local');
       if(this.localTrail && Array.isArray(this.localTrail)) {
-        let g:Geometry= this.trailLocal.getGeometry();
-        (g as LineString).setCoordinates( mapifyCoords(c) );
+        const g:Geometry= this.trailLocal.getGeometry();
+        (g as LineString).setCoordinates( c );
+        //(g as LineString).setCoordinates( mapifyCoords(c) );
       }
     }
   }
 
   parseServerTrail() {
     if(!this.serverTrail) { return }
-    let c= fromLonLatArray(this.serverTrail);
-    let ca= [];
-    c.forEach( (t:Array<Coordinate>)=> { ca.push(mapifyCoords(t)) });
+    let ca = this.serverTrail.map( (t:Array<Coordinate>)=> { return fromLonLatArray(mapifyCoords(t)) });
+    //let c= fromLonLatArray(this.serverTrail);
+    //let c= fromLonLatArray(this.serverTrail);
+    //let ca= [];
+    //c.forEach( (t:Array<Coordinate>)=> { ca.push(mapifyCoords(t)) });
     if(!this.trailServer) { // create feature
       this.trailServer= new Feature( new MultiLineString( ca ) );
       this.trailServer.setId('trail.self.server');
@@ -151,7 +156,7 @@ export class VesselTrailComponent implements OnInit, OnDestroy, OnChanges {
     else { //update feature
       this.trailServer= this.source.getFeatureById('trail.self.server');
       if(this.serverTrail && Array.isArray(this.serverTrail)) {
-        let g:Geometry= this.trailServer.getGeometry();
+        const g:Geometry= this.trailServer.getGeometry();
         (g as MultiLineString).setCoordinates(ca);
       }
     }