+++ src/app/modules/settings/settings-dialog.ts
@@ -103,4 +103,6 @@ export class SettingsDialog implements OnInit {
         this.facade.applySettings();
     }
 
+    noSort() { return 0 }
+
 }